+++ src/Config/ConfigProvider.php
@@ -19,39 +19,72 @@ use Soluble\MediaTools\VideoThumbService;
 
 class ConfigProvider
 {
+    /**
+     * @return array<string, array<string,array<string,mixed>>>
+     */
     public function __invoke(): array
     {
         return [
             'services' => [
                 'config' => $this->getDefaultConfiguration(),
             ],
-            'dependencies' => $this->getDependencies(),
+            'dependencies' => $this->getDependencies()
         ];
     }
 
+    /**
+     * @return array<string, array<string,string>>
+     */
     public function getDependencies(): array
     {
         return [
-            'aliases' => [
-                VideoConversionService::class => ConversionServiceInterface::class,
-                VideoInfoService::class       => InfoServiceInterface::class,
-                VideoDetectionService::class  => DetectionServiceInterface::class,
-                VideoThumbService::class      => ThumbServiceInterface::class,
-            ],
-            'factories' => [
-                // Configuration array for ffmpeg and ffprobe
-                FFMpegConfig::class   => FFMpegConfigFactory::class,
-                FFProbeConfig::class  => FFProbeConfigFactory::class,
-
-                // Services classes
-                ConversionServiceInterface::class => ConversionServiceFactory::class,
-                InfoServiceInterface::class       => InfoServiceFactory::class,
-                DetectionServiceInterface::class  => DetectionServiceFactory::class,
-                ThumbServiceInterface::class      => ThumbServiceFactory::class,
-            ],
+            'aliases' => $this->getAliases(),
+            'factories' => $this->getFactories()
         ];
     }
 
+    /**
+     * Return concrete implementation aliases if needed
+     *
+     * @return array<string, string>
+     */
+    public function getAliases(): array {
+
+        return [
+            // Configuration holders
+            FFMpegConfig::class           => FFMpegConfigInterface::class,
+            FFProbeConfig::class          => FFProbeConfigInterface::class,
+
+            // Services
+            VideoConversionService::class => ConversionServiceInterface::class,
+            VideoInfoService::class       => InfoServiceInterface::class,
+            VideoDetectionService::class  => DetectionServiceInterface::class,
+            VideoThumbService::class      => ThumbServiceInterface::class,
+        ];
+    }
+
+    /**
+     * Return interface based factories
+     *
+     * @return array<string, string>
+     */
+    public function getFactories(): array {
+        return [
+            // Configuration holders
+            FFMpegConfigInterface::class   => FFMpegConfigFactory::class,
+            FFProbeConfigInterface::class  => FFProbeConfigFactory::class,
+
+            // Services classes
+            ConversionServiceInterface::class => ConversionServiceFactory::class,
+            InfoServiceInterface::class       => InfoServiceFactory::class,
+            DetectionServiceInterface::class  => DetectionServiceFactory::class,
+            ThumbServiceInterface::class      => ThumbServiceFactory::class,
+        ];
+    }
+
+    /**
+     * @return array<string, array<string,mixed>>
+     */
     public function getDefaultConfiguration(): array
     {
         return require dirname(__DIR__, 2) . '/config/soluble-mediatools.global.php';