+++ src/Config/FFMpegConfig.php
@@ -4,14 +4,8 @@ declare(strict_types=1);
 
 namespace Soluble\MediaTools\Config;
 
-class FFMpegConfig
+class FFMpegConfig implements FFMpegConfigInterface
 {
-    public const DEFAULT_BINARY       = 'ffmpeg';
-    public const DEFAULT_THREADS      = null;
-    public const DEFAULT_TIMEOUT      = null;
-    public const DEFAULT_IDLE_TIMEOUT = null;
-    public const DEFAULT_ENV          = [];
-
     /** @var string */
     protected $binary;