+++ src/Video/InfoServiceFactory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Soluble\MediaTools\Video;
 
 use Psr\Container\ContainerInterface;
-use Soluble\MediaTools\Config\FFProbeConfig;
+use Soluble\MediaTools\Config\FFProbeConfigInterface;
 use Soluble\MediaTools\VideoInfoService;
 
 class InfoServiceFactory
@@ -13,7 +13,7 @@ class InfoServiceFactory
     public function __invoke(ContainerInterface $container): InfoServiceInterface
     {
         return new VideoInfoService(
-            $container->get(FFProbeConfig::class)
+            $container->get(FFProbeConfigInterface::class)
         );
     }
 }